% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt.R
\name{crypt}
\alias{crypt}
\title{Create crypted passwords}
\usage{
crypt(plaintext, salt = NULL)
}
\arguments{
\item{plaintext}{Plain text to be hashed}

\item{salt}{A two-character string used as "salt" for the hash.  If `NULL`, salt is randomly generated.}
}
\description{
WeBWorK passwords must be crypted with a 1-way hash. `crypt()` is a wrapper around the suggested perl code
for performing the crypting.
}
\examples{
crypt("plaintext")
crypt("plaintext", salt = "N7")
crypt("plaintext", salt = "Xa")
crypt(LETTERS)
}
