% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradebook.R
\name{ww_read_gradebook}
\alias{ww_read_gradebook}
\title{Ingest WeBWorK gradebook}
\usage{
ww_read_gradebook(path)
}
\arguments{
\item{path}{Path to gradebook file exported from WeBWorK}
}
\description{
Ingest WeBWorK gradebook
}
