% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt.R
\name{crypt1}
\alias{crypt1}
\title{Create crypted passwords}
\usage{
crypt1(plaintext, salt = NULL)
}
\arguments{
\item{plaintext}{Plain text to be hashed}

\item{salt}{Currently ignored.  Salt is generated randomly.}
}
\description{
Create cypted passwords for use with WeBWorK class lists.
}
