% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{ww_roster}
\alias{ww_roster}
\title{Roster utilities}
\usage{
ww_roster(email = "", user = sapply(stringr::str_split(email, "@"),
  function(x) x[1]), id = "", last_name = "", first_name = "",
  password = id, permission = 0, status = "current", comment = "",
  section = "", recitation = "")
}
\arguments{
\item{email}{Email address (may be left blank).}

\item{user}{User IDs (required).}

\item{id}{Student IDs (may be left blank).}

\item{last_name}{Last name (may be left blank).}

\item{first_name}{First name (may be left blank).}

\item{password}{Passwords (defaults to `student_id`).}

\item{permission}{Permission code: Use -5 for guest, 0 for student, 2 for login proctor, 3for grade proctor, 5 for teaching assistant, 10 for professor.}

\item{status}{One of `"current"`, `"enrolled"`, `"audit"`, `"drop"`.  (Can use c/a/d.)}

\item{comment}{A comment (may be left blank).}

\item{section}{Section desriptor (may be left blank).}

\item{recitation}{Recitaiton descriptor (may be left blank).}
}
\description{
Roster utilities
}
